generate mmt stuff in semantic tables
